requirements.txt
eth-brownie==1.19.2
git+https://github.com/deepsp94/ape-safe.git@contract
python-dotenv